prof-jd-2_0-env-hw7 после перевода на Maps
