+++ src/main/java/dev/ji/employeenewnew/service/EmployeeServiceCollections.java
@@ -1,15 +1,15 @@
-package dev.ji.employeenewnew.service;
-
-import dev.ji.employeenewnew.model.Employee;
-
-import java.util.Set;
-
-public interface EmployeeServiceCollections {
-    String addEmployee(String firstName, String lastName);
-
-    String removeEmployee(String firstName, String lastName);
-
-    Employee findEmployee(String firstName, String lastName);
-
-    Set<Employee> getAllEmployees();
-}
+//package dev.ji.employeenewnew.service;
+//
+//import dev.ji.employeenewnew.model.Employee;
+//
+//import java.util.Set;
+//
+//public interface EmployeeServiceCollections {
+//    String addEmployee(String firstName, String lastName);
+//
+//    String removeEmployee(String firstName, String lastName);
+//
+//    Employee findEmployee(String firstName, String lastName);
+//
+//    Set<Employee> getAllEmployees();
+//}
